requirements.txt
Babel==2.14.0
docxcompose==1.4.0
docxtpl==0.16.8
et-xmlfile==1.1.0
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
pdf2image==1.17.0
pillow==10.3.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
pywin32==306
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
